requirements.txt
robotframework>=6.0
robotframework-browser>=17.0
python-dotenv>=1.0
openpyxl>=3.1
pyyaml>=6.0
